Added view for comments, Changed order value from int to double, Changed option from bool to double for empty detection

+++ LoggingApp/LoggingApp/Models/FormModels.swift
@@ -55,7 +55,7 @@ class FormTemplate {
                         stepInLogbookItem.order = Int32(numberOfStep)
                         for (numberOfQuestion, question) in step.questionTitles.enumerated() {
                             let questionInStep = QuestionMO(context: context)
-                            questionInStep.order = Int32(numberOfQuestion)
+                            questionInStep.order = Double(numberOfQuestion)
                             questionInStep.title = question
                             questionInStep.addToStep(stepInLogbookItem)
                         }
